Corrections and new features
As it has been tested, some corrections are made in the code.

The Generator now uses one of the Adapters available (only CSV for
now). It now do NOT require an index in the CSV definitions (the column
must exist but can be empty) replacing it by the default language
value. This way, you can just write "translate('my english string')" to
use translations.